Code refactoring on moving models
